requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
absl-py==1.3.0
    # via
    #   chex
    #   jaxopt
    #   optax
antlr4-python3-runtime==4.9.3
    # via
    #   hydra-core
    #   omegaconf
chex==0.1.5
    # via optax
contourpy==1.0.6
    # via matplotlib
cvxpy==1.2.2
    # via -r requirements.in
cycler==0.11.0
    # via matplotlib
dm-tree==0.1.7
    # via chex
ecos==2.0.10
    # via cvxpy
fonttools==4.38.0
    # via matplotlib
hydra-core==1.2.0
    # via -r requirements.in
jax==0.3.25
    # via
    #   -r requirements.in
    #   chex
    #   jaxopt
    #   optax
jaxlib==0.3.25
    # via
    #   -r requirements.in
    #   chex
    #   jaxopt
    #   optax
jaxopt==0.5.5
    # via -r requirements.in
joblib==1.2.0
    # via scikit-learn
kiwisolver==1.4.4
    # via matplotlib
matplotlib==3.6.2
    # via
    #   -r requirements.in
    #   jaxopt
numpy==1.23.5
    # via
    #   -r requirements.in
    #   chex
    #   contourpy
    #   cvxpy
    #   ecos
    #   jax
    #   jaxlib
    #   jaxopt
    #   matplotlib
    #   opt-einsum
    #   optax
    #   osqp
    #   pandas
    #   scikit-learn
    #   scipy
    #   scs
omegaconf==2.2.3
    # via hydra-core
opt-einsum==3.3.0
    # via jax
optax==0.1.4
    # via -r requirements.in
osqp==0.6.2.post8
    # via cvxpy
packaging==21.3
    # via
    #   hydra-core
    #   matplotlib
pandas==1.5.2
    # via -r requirements.in
pillow==9.3.0
    # via matplotlib
pyparsing==3.0.9
    # via
    #   matplotlib
    #   packaging
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
pytz==2022.6
    # via pandas
pyyaml==6.0
    # via omegaconf
qdldl==0.1.5.post2
    # via osqp
scikit-learn==1.1.3
    # via -r requirements.in
scipy==1.9.3
    # via
    #   -r requirements.in
    #   cvxpy
    #   ecos
    #   jax
    #   jaxlib
    #   jaxopt
    #   osqp
    #   scikit-learn
    #   scs
scs==3.2.2
    # via cvxpy
six==1.16.0
    # via python-dateutil
sklearn==0.0.post1
    # via -r requirements.in
threadpoolctl==3.1.0
    # via scikit-learn
toolz==0.12.0
    # via chex
typing-extensions==4.4.0
    # via
    #   jax
    #   optax
